Expand with permutations of sara Am
Lines with a single sara Am are duplicated and in the first line the sara Am is replaced with a nikhit and sara A; lines with 2 sara Ams are quadrupled and all 4 permutations of sara Am / nikhit + sara A are applied in th_TH.UTF-8 collating order. Purpose: correctly recognize and break up mangled texts.